feat: Switched from JavaScript to TypeScript.
