+++ css/statShop.css
@@ -1,0 +1,23 @@
+#stat-shop-interface {
+    position: absolute;
+    margin: auto;
+    display: none;
+    left: 0;
+    right: 0;
+    top: 10%;
+    width: 80%;
+    height: 80%;
+    border-radius: 5rem;
+    background-image: url("https://www.shutterstock.com/shutterstock/videos/3482446057/thumb/8.jpg?ip=x480");
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center;
+}
+
+#stat-shop-pokemonteam {
+    display: none;
+    width: 80%;
+    height: 80%;
+    background-color: red;
+    position: absolute;
+}
