+++ package.json
@@ -1,0 +1,21 @@
+{
+  "name": "pokemon-game",
+  "version": "1.0.0",
+  "description": "My pokemon game :)",
+  "main": "index.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc --build && vite build",
+    "clean": "tsc --build --clean",
+    "dev": "vite"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "typescript": "^5.6.3",
+    "vite": "^5.4.11"
+  },
+  "devDependencies": {
+    "@types/node": "^22.9.0"
+  }
+}
