--- scripts/constants.js
@@ -1,1 +1,0 @@
-var DEV_MODE = true;
