+++ scripts/constants.ts
@@ -1,0 +1,1 @@
+export const DEV_MODE = true;
