+++ scripts/healShop.ts
@@ -1,0 +1,20 @@
+import { updateAllyPokemon } from "./battleSceneModule";
+
+
+export const healPokemonTeam = (pokemonList) => {
+    if (!pokemonList) {
+        return;
+    }
+    for (let i = 0; i < pokemonList.length; i++) {
+        pokemonList[i].currentHp = pokemonList[i].pokemonType.health(pokemonList[i].level)
+    }
+    updateAllyPokemon();
+}
+
+export const healPokemon = (pokemonIndividual, healAmount: number) => {
+    const maxHp = pokemonIndividual.pokemonType.health(pokemonIndividual.level);
+    const newHp = pokemonIndividual.currentHp + healAmount;
+    const actualNewHp = Math.min(maxHp, newHp);
+
+    pokemonIndividual.currentHp = actualNewHp;
+};
