--- scripts/heldItemsModule.js
@@ -1,15 +1,0 @@
-const heldItemsModule = (function () {
-
-    const ALL_ITEMS = {
-        expshare: {
-            name: "Exp. Share"
-        },
-        expall: {
-            name: "Exp. All"
-        }
-    }
-
-    return {
-        ALL_ITEMS
-    }
-})()
