+++ scripts/heldItemsModule.ts
@@ -1,0 +1,8 @@
+export const ALL_ITEMS = {
+    expshare: {
+        name: "Exp. Share"
+    },
+    expall: {
+        name: "Exp. All"
+    }
+}
