+++ scripts/pokemonModule.ts
@@ -1,0 +1,873 @@
+
+function healthGenerator(level: number): number {
+  return 10 + level + Math.floor(level * 0.01 * 2 * this.hp);
+};
+
+//HP = floor(0.01 x (2 x Base + IV + floor(0.25 x EV)) x Level) + Level + 10
+//list used for wild encounter fights.
+
+
+
+type Move = [string, number];
+
+export type PokemonType = {
+  id: string;
+  name: string;
+  health: (level: number) => number;
+  moves: Move[];
+  hp: number;
+  attack: number;
+  defense: number;
+  spatk: number;
+  spdef: number;
+  speed: number;
+  baseExp: number;
+  basePrice: number;
+  targetEvolution?: string;
+  evolveLevel?: number;
+}
+
+export const ALL_POKEMON: Record<string, PokemonType> = {
+  bulbasaur: {
+    id: "0001-bulbasaur",
+    name: "Bulbasaur",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1],
+      ["vinewhip", 6],
+      ["razorleaf", 12],
+      ["seedbomb", 18],
+      ["takedown", 21],
+      ["powerwhip", 33],
+      ["solarbeam", 36],
+    ],
+    hp: 45,
+    attack: 49,
+    defense: 49,
+    spatk: 65,
+    spdef: 65,
+    speed: 45,
+    baseExp: 100,
+    basePrice: 50000,
+    targetEvolution: "Ivysaur",
+    evolveLevel: 16,
+  },
+  ivysaur: {
+    id: "0002-ivysaur",
+    name: "Ivysaur",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1],
+      ["vinewhip", 6],
+      ["razorleaf", 12],
+      ["seedbomb", 18],
+      ["takedown", 21],
+      ["powerwhip", 33],
+      ["solarbeam", 36],
+    ],
+    hp: 60,
+    attack: 62,
+    defense: 53,
+    spatk: 80,
+    spdef: 80,
+    speed: 60,
+    baseExp: 100,
+    basePrice: 50000,
+    targetEvolution: "Venusaur",
+    evolveLevel: 36,
+  },
+  venusaur: {
+    id: "0003-venusaur",
+    name: "Venusaur",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1],
+      ["vinewhip", 6],
+      ["razorleaf", 12],
+      ["seedbomb", 18],
+      ["takedown", 21],
+      ["powerwhip", 33],
+      ["solarbeam", 36],
+    ],
+    hp: 80,
+    attack: 100,
+    defense: 123,
+    spatk: 122,
+    spdef: 120,
+    speed: 80,
+    baseExp: 100,
+    basePrice: 500000,
+  },
+  charmander: {
+    id: "0004-charmander",
+    name: "Charmander",
+    health: healthGenerator,
+    moves: [
+      ["scratch", 1],
+      ["ember", 6],
+      ["dragonbreath", 12],
+      ["firefang", 17],
+      ["slash", 20],
+      ["flamethrower", 24],
+      ["inferno", 36],
+      ["flareblitz", 40],
+    ],
+    hp: 39,
+    attack: 52,
+    defense: 43,
+    spatk: 60,
+    spdef: 50,
+    speed: 65,
+    baseExp: 100,
+    basePrice: 50000,
+    targetEvolution: "Charmeleon",
+    evolveLevel: 16,
+  },
+  charmeleon: {
+    id: "0005-charmeleon",
+    name: "Charmeleon",
+    health: healthGenerator,
+    moves: [
+      ["scratch", 1],
+      ["ember", 6],
+      ["dragonbreath", 12],
+      ["firefang", 17],
+      ["slash", 20],
+      ["flamethrower", 24],
+      ["inferno", 36],
+      ["flareblitz", 40],
+    ],
+    hp: 58,
+    attack: 64,
+    defense: 58,
+    spatk: 80,
+    spdef: 65,
+    speed: 80,
+    baseExp: 100,
+    basePrice: 50000,
+    targetEvolution: "Charizard",
+    evolveLevel: 36,
+  },
+  charizard: {
+    id: "0006-charizard",
+    name: "Charizard",
+    health: healthGenerator,
+    moves: [
+      ["scratch", 1],
+      ["ember", 6],
+      ["dragonbreath", 12],
+      ["firefang", 17],
+      ["slash", 20],
+      ["flamethrower", 24],
+      ["inferno", 36],
+      ["flareblitz", 40],
+    ],
+    hp: 78,
+    attack: 84,
+    defense: 78,
+    spatk: 109,
+    spdef: 85,
+    speed: 100,
+    baseExp: 100,
+    basePrice: 500000,
+  },
+  squirtle: {
+    id: "0007-squirtle",
+    name: "Squirtle",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1],
+      ["watergun", 6],
+      ["bite", 12],
+      ["waterpulse", 15],
+      ["aqua tail", 24],
+      ["hydropump", 33],
+      ["wavecrash", 36],
+    ],
+    hp: 44,
+    attack: 48,
+    defense: 65,
+    spatk: 50,
+    spdef: 64,
+    speed: 43,
+    baseExp: 100,
+    basePrice: 50000,
+    targetEvolution: "Wartortle",
+    evolveLevel: 16,
+  },
+  wartortle: {
+    id: "0008-wartortle",
+    name: "Wartortle",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1],
+      ["watergun", 6],
+      ["bite", 12],
+      ["waterpulse", 15],
+      ["aqua tail", 24],
+      ["hydropump", 33],
+      ["wavecrash", 36],
+    ],
+    hp: 59,
+    attack: 63,
+    defense: 80,
+    spatk: 65,
+    spdef: 80,
+    speed: 58,
+    baseExp: 100,
+    basePrice: 50000,
+    targetEvolution: "Blastoise",
+    evolveLevel: 36,
+  },
+  blastoise: {
+    id: "0009-blastoise",
+    name: "Blastoise",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1],
+      ["watergun", 6],
+      ["bite", 12],
+      ["waterpulse", 15],
+      ["aqua tail", 24],
+      ["hydropump", 33],
+      ["wavecrash", 36],
+    ],
+    hp: 79,
+    attack: 83,
+    defense: 100,
+    spatk: 85,
+    spdef: 100,
+    speed: 78,
+    baseExp: 100,
+    basePrice: 500000,
+  },
+  butterfree: {
+    id: "0012-butterfree",
+    name: "Butterfree",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 60,
+    attack: 45,
+    defense: 50,
+    spatk: 90,
+    spdef: 80,
+    speed: 70,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  rattata: {
+    id: "0019-rattata",
+    name: "Rattata",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 30,
+    attack: 56,
+    defense: 35,
+    spatk: 25,
+    spdef: 35,
+    speed: 72,
+    baseExp: 51,
+    basePrice: 1000,
+  },
+  raticate: {
+    id: "0020-raticate",
+    name: "Raticate",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 55,
+    attack: 81,
+    defense: 60,
+    spatk: 50,
+    spdef: 70,
+    speed: 97,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  pidgey: {
+    id: "0016-pidgey",
+    name: "Pidgey",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 45,
+    defense: 40,
+    spatk: 35,
+    spdef: 35,
+    speed: 56,
+    baseExp: 55,
+    basePrice: 1000,
+  },
+  pidgeotto: {
+    id: "0017-pidgeotto",
+    name: "Pidgeotto",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 63,
+    attack: 60,
+    defense: 55,
+    spatk: 50,
+    spdef: 50,
+    speed: 71,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  pidgeott: {
+    id: "0018-pidgeott",
+    name: "Pidgeotto",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 63,
+    attack: 60,
+    defense: 55,
+    spatk: 50,
+    spdef: 50,
+    speed: 71,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  spearow: {
+    id: "0021-spearow",
+    name: "Spearow",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 60,
+    defense: 30,
+    spatk: 31,
+    spdef: 31,
+    speed: 70,
+    baseExp: 100,
+    basePrice: 1000,
+  },
+  fearow: {
+    id: "0022-fearow",
+    name: "Fearow",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 65,
+    attack: 90,
+    defense: 65,
+    spatk: 61,
+    spdef: 61,
+    speed: 100,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  ekans: {
+    id: "0023-ekans",
+    name: "Ekans",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 35,
+    attack: 60,
+    defense: 44,
+    spatk: 40,
+    spdef: 54,
+    speed: 55,
+    baseExp: 100,
+    basePrice: 1000,
+  },
+  arbok: {
+    id: "0024-arbok",
+    name: "Arbok",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 60,
+    attack: 95,
+    defense: 69,
+    spatk: 65,
+    spdef: 79,
+    speed: 80,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  pikachu: {
+    id: "0025-pikachu",
+    name: "Pikachu",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 35,
+    attack: 55,
+    defense: 40,
+    spatk: 50,
+    spdef: 50,
+    speed: 90,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  raichu: {
+    id: "0026-raichu",
+    name: "Raichu",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 60,
+    attack: 90,
+    defense: 55,
+    spatk: 90,
+    spdef: 80,
+    speed: 110,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  sandshrew: {
+    id: "0027-sandshrew",
+    name: "Sandshrew",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 50,
+    attack: 75,
+    defense: 85,
+    spatk: 20,
+    spdef: 30,
+    speed: 40,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  nidoranfemale: {
+    id: "0029-nidoranfemale",
+    name: "Nidoran",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 55,
+    attack: 47,
+    defense: 52,
+    spatk: 40,
+    spdef: 40,
+    speed: 41,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  nidoranmale: {
+    id: "0032-nidoranmale",
+    name: "Nidoran",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 46,
+    attack: 57,
+    defense: 40,
+    spatk: 40,
+    spdef: 40,
+    speed: 50,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  vulpix: {
+    id: "0037-vulpix",
+    name: "Vulpix",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 38,
+    attack: 41,
+    defense: 40,
+    spatk: 50,
+    spdef: 65,
+    speed: 65,
+    baseExp: 100,
+    basePrice: 50000,
+  },
+  jigglypuff: {
+    id: "0039-jigglypuff",
+    name: "Jigglypuff",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 115,
+    attack: 45,
+    defense: 20,
+    spatk: 45,
+    spdef: 25,
+    speed: 20,
+    baseExp: 100,
+    basePrice: 2000,
+  },
+  oddish: {
+    id: "0043-oddish",
+    name: "Oddish",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 45,
+    attack: 50,
+    defense: 55,
+    spatk: 75,
+    spdef: 65,
+    speed: 30,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  meowth: {
+    id: "0052-meowth",
+    name: "Meowth",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 45,
+    defense: 35,
+    spatk: 40,
+    spdef: 40,
+    speed: 90,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  mankey: {
+    id: "0056-mankey",
+    name: "Mankey",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 80,
+    defense: 35,
+    spatk: 35,
+    spdef: 45,
+    speed: 70,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  growlithe: {
+    id: "0058-growlithe",
+    name: "Growlithe",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 55,
+    attack: 70,
+    defense: 45,
+    spatk: 70,
+    spdef: 50,
+    speed: 60,
+    baseExp: 100,
+    basePrice: 50000,
+  },
+  abra: {
+    id: "0063-abra",
+    name: "Abra",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 25,
+    attack: 20,
+    defense: 15,
+    spatk: 105,
+    spdef: 55,
+    speed: 90,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  kadabra: {
+    id: "0064-kadabra",
+    name: "Kadabra",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 35,
+    defense: 30,
+    spatk: 120,
+    spdef: 70,
+    speed: 105,
+    baseExp: 100,
+    basePrice: 50000,
+  },
+  machop: {
+    id: "0066-machop",
+    name: "Machop",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 70,
+    attack: 80,
+    defense: 50,
+    spatk: 35,
+    spdef: 35,
+    speed: 35,
+    baseExp: 100,
+    basePrice: 25000,
+  },
+  bellsprout: {
+    id: "0069-bellsprout",
+    name: "Bellsprout",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 50,
+    attack: 75,
+    defense: 35,
+    spatk: 70,
+    spdef: 30,
+    speed: 40,
+    baseExp: 100,
+    basePrice: 15000,
+  },
+  magnemite: {
+    id: "0081-magnemite",
+    name: "Magnemite",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 25,
+    attack: 35,
+    defense: 70,
+    spatk: 95,
+    spdef: 55,
+    speed: 45,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  voltorb: {
+    id: "0100-voltorb",
+    name: "Voltorb",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 30,
+    defense: 50,
+    spatk: 55,
+    spdef: 55,
+    speed: 100,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  mew: {
+    id: "0151-mew",
+    name: "Mew",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 100,
+    attack: 100,
+    defense: 100,
+    spatk: 100,
+    spdef: 100,
+    speed: 100,
+    baseExp: 100,
+    basePrice: 1000000,
+  },
+  eevee: {
+    id: "0133-eevee",
+    name: "Eevee",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 55,
+    attack: 55,
+    defense: 50,
+    spatk: 45,
+    spdef: 65,
+    speed: 55,
+    baseExp: 100,
+    basePrice: 25000,
+  },
+  mimikyu: {
+    id: "0778-mimikyu",
+    name: "Mimikyu",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 55,
+    attack: 90,
+    defense: 80,
+    spatk: 50,
+    spdef: 105,
+    speed: 96,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  rockruff: {
+    id: "0744-rockruff",
+    name: "Rockruff",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 45,
+    attack: 65,
+    defense: 40,
+    spatk: 30,
+    spdef: 40,
+    speed: 60,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  snorlax: {
+    id: "0143-snorlax",
+    name: "Snorlax",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 160,
+    attack: 110,
+    defense: 65,
+    spatk: 65,
+    spdef: 110,
+    speed: 30,
+    baseExp: 100,
+    basePrice: 250000,
+  },
+
+  beedrill: {
+    id: "0015-beedrill",
+    name: "Beedrill",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 65,
+    attack: 80,
+    defense: 40,
+    spatk: 45,
+    spdef: 80,
+    speed: 75,
+    baseExp: 100,
+    basePrice: 5000,
+  },
+  kakuna: {
+    id: "0014-kakuna",
+    name: "Kakuna",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 45,
+    attack: 25,
+    defense: 50,
+    spatk: 25,
+    spdef: 25,
+    speed: 35,
+    baseExp: 100,
+    basePrice: 2000,
+  },
+  weedle: {
+    id: "0013-weedle",
+    name: "Weedle",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 40,
+    attack: 35,
+    defense: 30,
+    spatk: 20,
+    spdef: 20,
+    speed: 50,
+    baseExp: 100,
+    basePrice: 500,
+  },
+  caterpie: {
+    id: "0010-caterpie",
+    name: "Caterpie",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 45,
+    attack: 30,
+    defense: 35,
+    spatk: 20,
+    spdef: 20,
+    speed: 45,
+    baseExp: 100,
+    basePrice: 500,
+  },
+  metapod: {
+    id: "0011-metapod",
+    name: "Metapod",
+    health: healthGenerator,
+    moves: [
+      ["tackle", 1]
+    ],
+    hp: 50,
+    attack: 20,
+    defense: 5,
+    spatk: 25,
+    spdef: 25,
+    speed: 30,
+    baseExp: 100,
+    basePrice: 2000,
+  },
+};
+
+export const POKEMON_FOR_SALE: PokemonType[] = [
+  "0001-bulbasaur",
+  "0004-charmander",
+  "0007-squirtle",
+  "0010-caterpie",
+  "0011-metapod",
+  "0013-weedle",
+  "0016-pidgey",
+  "0021-spearow",
+  "0019-rattata",
+  "0023-ekans",
+  "0025-pikachu",
+  "0027-sandshrew",
+  "0029-nidoranfemale",
+  "0032-nidoranmale",
+  "0037-vulpix",
+  "0039-jigglypuff",
+  "0043-oddish",
+  "0052-meowth",
+  "0056-mankey",
+  "0058-growlithe",
+  "0063-abra",
+  "0066-machop",
+  "0069-bellsprout",
+  "0081-magnemite",
+  "0100-voltorb",
+  "0133-eevee",
+  "0143-snorlax",
+  "0151-mew",
+  "0744-rockruff",
+  "0778-mimikyu",
+].map(id => Object.values(ALL_POKEMON).find(pokemonType => pokemonType.id == id))
+  .filter((pokemon): pokemon is PokemonType => !!pokemon);
+
+export const ALL_POKEMON_ARRAY = Object.values(ALL_POKEMON);
