+++ scripts/sharedData.ts
@@ -1,0 +1,18 @@
+
+let pokeCurrency = 0;
+
+export const playerPokemonList = [];
+
+export const getPokeCurrency = () => pokeCurrency;
+
+export const setPokeCurrency = (amount: number) => {
+    pokeCurrency = amount
+}
+
+export const addPokeCurrency = (amount: number) => {
+    pokeCurrency += amount;
+}
+
+export const removePokeCurrency = (amount: number) => {
+    pokeCurrency -= amount;
+}
