+++ scripts/trainerModule.ts
@@ -1,0 +1,15 @@
+import { ALL_POKEMON } from "./pokemonModule";
+
+export const TRAINERS = {
+    Anna: {
+        name: "Anna",
+        pokemonTeam: [
+            ALL_POKEMON.pidgey,
+            ALL_POKEMON.rattata,
+            ALL_POKEMON.mankey,
+            ALL_POKEMON.jigglypuff,
+            ALL_POKEMON.machop,
+            ALL_POKEMON.charmander,
+        ]
+    }
+}
